fix for new rust

+++ src/cpus.rs
@@ -4,7 +4,7 @@ use _macros;
 
 pub fn num_cpus() -> uint {
     unsafe {
-        return rust_get_num_cpus();
+        return rust_get_num_cpus() as uint;
     }
     extern {
         fn rust_get_num_cpus() -> libc::uintptr_t;
